Stinking Cloud (#203)
* Add Stinking Cloud edits

Vanilla spell blocks actions on a failed save, but the save is rolled as soon as you enter the cloud. RAW, it should allow a character to pass through it without needing to save, only requiring a save vs. the effect when starting a turn within the cloud.

* Add files via upload

* Add files via upload

* Add new Stinking Cloud descriptions

* Update ModOptions.json

Added Stinking Cloud to Spell.list options

* Adding changes to mod's description

---------

Co-authored-by: FelipeRenault <f.lemos.renault@gmail.com>